# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API for Ferme Familiale - Gestion de la ferme avicole
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bandes:
    get:
      operationId: listBandes
      tags: [bandes]
      summary: Liste des bandes de poulets
      responses:
        "200":
          description: Bandes de poulets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bande"
    post:
      operationId: createBande
      tags: [bandes]
      summary: Créer une bande
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBandeBody"
      responses:
        "201":
          description: Bande créée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bande"
  /bandes/{id}:
    get:
      operationId: getBande
      tags: [bandes]
      summary: Détails d'une bande
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Détails de la bande
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BandeDetail"
    put:
      operationId: updateBande
      tags: [bandes]
      summary: Mettre à jour une bande
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBandeBody"
      responses:
        "200":
          description: Bande mise à jour
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bande"
    delete:
      operationId: deleteBande
      tags: [bandes]
      summary: Supprimer une bande
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bande supprimée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /bandes/{id}/depenses:
    get:
      operationId: listBandeDepenses
      tags: [bandes]
      summary: Liste des dépenses d'une bande
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dépenses de la bande
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BandeDepense"
    post:
      operationId: createBandeDepense
      tags: [bandes]
      summary: Créer une dépense de bande
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBandeDepenseBody"
      responses:
        "201":
          description: Dépense créée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BandeDepense"
  /bandes/{id}/depenses/{depenseId}:
    put:
      operationId: updateBandeDepense
      tags: [bandes]
      summary: Mettre à jour une dépense de bande
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: depenseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBandeDepenseBody"
      responses:
        "200":
          description: Dépense mise à jour
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BandeDepense"
    delete:
      operationId: deleteBandeDepense
      tags: [bandes]
      summary: Supprimer une dépense de bande
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: depenseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dépense supprimée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /bandes/{id}/ventes:
    get:
      operationId: listBandeVentes
      tags: [bandes]
      summary: Liste des ventes d'une bande
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ventes de la bande
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BandeVente"
    post:
      operationId: createBandeVente
      tags: [bandes]
      summary: Créer une vente de bande
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBandeVenteBody"
      responses:
        "201":
          description: Vente créée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BandeVente"
  /bandes/{id}/ventes/{venteId}:
    put:
      operationId: updateBandeVente
      tags: [bandes]
      summary: Mettre à jour une vente de bande
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: venteId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBandeVenteBody"
      responses:
        "200":
          description: Vente mise à jour
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BandeVente"
    delete:
      operationId: deleteBandeVente
      tags: [bandes]
      summary: Supprimer une vente de bande
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: venteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vente supprimée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /bandes/{id}/charges-fixes:
    get:
      operationId: getBandeChargesFixe
      tags: [bandes]
      summary: Charges fixes d'une bande
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Charges fixes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChargesFixe"
    put:
      operationId: updateBandeChargesFixe
      tags: [bandes]
      summary: Mettre à jour les charges fixes d'une bande
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateChargesFixeBody"
      responses:
        "200":
          description: Charges fixes mises à jour
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChargesFixe"
  /bandes/{id}/depenses-vente:
    get:
      operationId: listBandeDepensesVente
      tags: [bandes]
      summary: Dépenses liées à la vente d'une bande
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dépenses de vente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DepenseVente"
    post:
      operationId: createBandeDepenseVente
      tags: [bandes]
      summary: Créer une dépense de vente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepenseVenteBody"
      responses:
        "201":
          description: Dépense de vente créée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepenseVente"
  /bandes/{id}/depenses-vente/{depenseId}:
    put:
      operationId: updateBandeDepenseVente
      tags: [bandes]
      summary: Mettre à jour une dépense de vente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: depenseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepenseVenteBody"
      responses:
        "200":
          description: Dépense de vente mise à jour
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepenseVente"
    delete:
      operationId: deleteBandeDepenseVente
      tags: [bandes]
      summary: Supprimer une dépense de vente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: depenseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Suppression réussie
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    Bande:
      type: object
      properties:
        id:
          type: integer
        numero:
          type: integer
        nom:
          type: string
        sujetsDepart:
          type: integer
        nombreDeces:
          type: integer
        valeurMaterielFixe:
          type: number
        statut:
          type: string
          enum: [active, terminee]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - numero
        - nom
        - sujetsDepart
        - nombreDeces
        - valeurMaterielFixe
        - statut
        - createdAt
    CreateBandeBody:
      type: object
      properties:
        nom:
          type: string
        sujetsDepart:
          type: integer
        nombreDeces:
          type: integer
        valeurMaterielFixe:
          type: number
        statut:
          type: string
          enum: [active, terminee]
      required:
        - nom
        - sujetsDepart
    BandeDetail:
      type: object
      properties:
        id:
          type: integer
        numero:
          type: integer
        nom:
          type: string
        sujetsDepart:
          type: integer
        nombreDeces:
          type: integer
        sujetsRestants:
          type: integer
        valeurMaterielFixe:
          type: number
        statut:
          type: string
          enum: [active, terminee]
        totalDepenses:
          type: number
        totalRecettes:
          type: number
        chargesFixesTotal:
          type: number
        beneficeNet:
          type: number
        beneficeNetSansCharges:
          type: number
        coutParSujet:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - numero
        - nom
        - sujetsDepart
        - nombreDeces
        - sujetsRestants
        - valeurMaterielFixe
        - statut
        - totalDepenses
        - totalRecettes
        - chargesFixesTotal
        - beneficeNet
        - beneficeNetSansCharges
        - coutParSujet
        - createdAt
    BandeDepense:
      type: object
      properties:
        id:
          type: integer
        bandeId:
          type: integer
        designation:
          type: string
        categorie:
          type: string
          enum: [poussins, aliments, concentre, medicaments, carburant, salaires, materiel_divers]
        quantite:
          type: number
        prixUnitaire:
          type: number
        montant:
          type: number
      required:
        - id
        - bandeId
        - designation
        - categorie
        - quantite
        - prixUnitaire
        - montant
    CreateBandeDepenseBody:
      type: object
      properties:
        designation:
          type: string
        categorie:
          type: string
          enum: [poussins, aliments, concentre, medicaments, carburant, salaires, materiel_divers]
        quantite:
          type: number
        prixUnitaire:
          type: number
      required:
        - designation
        - categorie
        - quantite
        - prixUnitaire
    BandeVente:
      type: object
      properties:
        id:
          type: integer
        bandeId:
          type: integer
        date:
          type: string
          format: date
        quantiteVendue:
          type: integer
        prixUnitaire:
          type: number
        montant:
          type: number
      required:
        - id
        - bandeId
        - date
        - quantiteVendue
        - prixUnitaire
        - montant
    CreateBandeVenteBody:
      type: object
      properties:
        date:
          type: string
          format: date
        quantiteVendue:
          type: integer
        prixUnitaire:
          type: number
      required:
        - date
        - quantiteVendue
        - prixUnitaire
    ChargesFixe:
      type: object
      properties:
        id:
          type: integer
        bandeId:
          type: integer
        valeurPerdueMateriel:
          type: number
        imprévus:
          type: number
        loyer:
          type: number
        total:
          type: number
      required:
        - id
        - bandeId
        - valeurPerdueMateriel
        - imprévus
        - loyer
        - total
    UpdateChargesFixeBody:
      type: object
      properties:
        loyer:
          type: number
    DepenseVente:
      type: object
      properties:
        id:
          type: integer
        bandeId:
          type: integer
        designation:
          type: string
        montant:
          type: number
      required:
        - id
        - bandeId
        - designation
        - montant
    CreateDepenseVenteBody:
      type: object
      properties:
        designation:
          type: string
        montant:
          type: number
      required:
        - designation
        - montant
...